Remove extraneous label_method and value_method.
simple_form, it turns out, is smart enough to inspect the object in the
collection to guess at the label_method and value_method. If it's an
array it guesses first and second map to label and value. If it's an
object, it looks for "common" label and value methods. These are
settable in the simple_forms initilizer (see config.collection_label_methods
aond config.collection_value_methods) but the default includes :name and
:id for the label and value methods. This works great with the
convention being used with our domain tables of having the value be the
id and the name be the option string.

Thus, this CL deletes a bunch of extra crufty lines.