Remove sci_type and use empty? idiom.
The inclusion of the patient.sci_type field into the _asia_fields
partial by using wrapped_form_for was causing a nested form attribute
that confused simple_forms breaking the sumbit button. It's unclear what
the right UI paradigm is for this field so moving back for now. This
reopens #260.