Make heroku deploy more robust to migrations
app:seed_fake uses factory girl to load data which relies on having model
code loaded. Because some of our migrations have been written with
dependencies on model code, this causes a race condition where a model
may not reflect the final state on an initialization from scratch.

Break apart the command with && to force rails to reload the model
classes after the migration.