Remove all references to models from migrations.
This breaks some data migrations, but frankly our migrations to date
are not correctly reversible anyways. Also the app has not gone into
production so who cares about anything beyond the schema.

Not accessing the model classes at all in the migrations is a best
practice and keeps a lot of other stuff working sanely.